ChatBot_carrefour/backend/app/requirements.txt
fastapi
uvicorn
pandas
requests
beautifulsoup4
python-dotenv
groq
